+++ internal/features/configuration/dto.go
@@ -7,7 +7,7 @@ import (
 // Request DTOs
 type CreateConfigurationRequest struct {
 	Configuration struct {
-		AccountID 						int `json:"account_id" validate:"required,gt=0"`
+		AccountID 						int64 `json:"account_id" validate:"required,gt=0"`
 		DefaultConfiguration 	*bool `json:"default_configuration" validate:"required"`
 		ConfigType 						string `json:"config_type" validate:"required,oneof=smtp in_app"`
 		ConfigurationData 		map[string]any `json:"configuration_data" validate:"required"`
@@ -16,8 +16,8 @@ type CreateConfigurationRequest struct {
 
 type UpdateConfigurationRequest struct {
 	Configuration struct {
-		ID 											int `json:"id" validate:"omitempty,gt=0"`
-		AccountID 							int `json:"account_id" validate:"omitempty,gt=0"`
+		ID 											int64 `json:"id" validate:"omitempty,gt=0"`
+		AccountID 							int64 `json:"account_id" validate:"omitempty,gt=0"`
 		DefaultConfiguration 		*bool `json:"default_configuration" validate:"omitempty,boolean"`
 		ConfigType 							string `json:"config_type" validate:"omitempty,oneof=smtp in_app"`
 		ConfigurationData 			map[string]any `json:"configuration_data" validate:"omitempty"`
@@ -25,8 +25,8 @@ type UpdateConfigurationRequest struct {
 }
 
 type ConfigurationPayload struct {
-	ID 										int `json:"id" validate:"omitempty,gt=0"`
-	AccountID 						int `json:"account_id" validate:"required,gt=0"`
+	ID 										int64 `json:"id" validate:"omitempty,gt=0"`
+	AccountID 						int64 `json:"account_id" validate:"required,gt=0"`
 	DefaultConfiguration 	*bool `json:"default_configuration" validate:"omitempty,boolean"`
 	ConfigType 						string `json:"config_type" validate:"omitempty,oneof=smtp in_app"`
 	ConfigurationData 		map[string]any `json:"configuration_data" validate:"omitempty"`
@@ -38,8 +38,8 @@ type ConfigurationRequest struct {
 
 // Response DTOs
 type CreateConfigurationResponse struct {
-	ID 										int `json:"id"`
-	AccountID 						int `json:"account_id"`
+	ID 										int64 `json:"id"`
+	AccountID 						int64 `json:"account_id"`
 	DefaultConfiguration 	bool `json:"default_configuration"`
 	ConfigType 						string `json:"config_type"`
 	ConfigurationData 		map[string]any `json:"configuration_data"`
