+++ internal/shared/dto/configuration.go
@@ -4,7 +4,7 @@ package dto
 
 type SMTPConfiguration struct {
 	Host     string `json:"host" validate:"required,min=1"`
-	Port     int    `json:"port" validate:"required,gt=0,lte=65535"`
+	Port     int64    `json:"port" validate:"required,gt=0,lte=65535"`
 	Username string `json:"username" validate:"required,min=1"`
 	Password string `json:"password" validate:"required,min=1"`
 	From     string `json:"from" validate:"required,email"`
